Initial Asteroid Spawning code

+++ Assets/Scripts/ScriptableObjectScripts/Asteroid_SO.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [CreateAssetMenu(menuName = "Scriptable Object/Asteroid_SO", fileName = "Asteroid")]
 public class Asteroid_SO : ScriptableObject
 {
-    public float initialSize = 3;
+    public int initialSize = 3;
     public float initialSpeed = 10;
     public GameObject[] asteroidPrefabs;
